requirement.txt
asgiref==3.3.4
attrs==21.2.0
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
classadd @ file:///C:/django/company/pure/classadd/dist/classadd-0.1.tar.gz
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
defusedxml==0.7.1
Django==3.2.4
django-allauth==0.44.0
django-rest-auth==0.9.5
djangorestframework==3.12.4
drf-yasg==1.20.0
gunicorn==20.1.0
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
jsonschema==3.2.0
MarkupSafe==2.0.1
micropython-ffilib==0.1.2
oauthlib==3.1.1
packaging==20.9
pycparser==2.20
PyJWT==2.1.0
pyparsing==2.4.7
pyrsistent==0.17.3
python3-openid==3.2.0
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
ruamel.yaml==0.17.10
ruamel.yaml.clib==0.2.4
six==1.16.0
sqlparse==0.4.1
swagger-spec-validator==2.7.3
uritemplate==3.0.1
urllib3==1.26.6
worker1 @ file:///C:/django/pure/worker1/dist/worker1-0.1.tar.gz
